app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastmcp>=3.2.0",
    "llama-index>=0.14.19",
    "llama-index-llms-google-genai>=0.9.1",
    "nicegui>=3.9.0",
    "pydantic>=2.12.5",
]
